# perf(pkgs-graph): speed up createPkgGraph when directory specifiers are present