--- a/.changeset/twelve-gifts-rescue.md
+++ b/.changeset/twelve-gifts-rescue.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/workspace.pkgs-graph": patch
+---
+
+Speed up createPkgGraph when directory specifiers are present