This speeds up the install of "with paths" test case (https://github.com/jakebailey/DefinitelyTyped/tree/blog-pnpm-workspaces-with-paths) from 49.7s to 14s.

Time spent in `createPkgMap` has been reduced from ~35s to ~200ms, matching the performance in #6287.

Fixes #6281 (for real this time)