@jakebailey 2023-03-29
Clearly I celebrated too early; Windows seems to be misbehaving. I'll take a look.